refactor(editor): Remove the NEXT env variables in the editor built instead of the web component built
