+++ packages/editor-web-component/vite.config.ts
@@ -6,45 +6,21 @@ import { Plugin, defineConfig } from 'vite'
 import cssInjectedByJsPlugin from 'vite-plugin-css-injected-by-js'
 import dts from 'vite-plugin-dts'
 import svgr from 'vite-plugin-svgr'
-import replace from '@rollup/plugin-replace'
 import * as acorn from 'acorn'
 import * as estraverse from 'estraverse'
 import * as escodegen from 'escodegen'
 import { type Program, type Node } from 'estree'
+import replace from '@rollup/plugin-replace'
 
 // https://vitejs.dev/guide/build.html#library-mode
 
-const js = (value: string) => JSON.stringify(value)
-
-const productionKeys = ['process.env.NODE_ENV', 'process.env.NEXT_PUBLIC_ENV']
-
-const notProvidedKeys = [
-  '__NEXT_I18N_SUPPORT',
-  '__NEXT_ROUTER_BASEPATH',
-  '__NEXT_CLIENT_ROUTER_D_FILTER',
-  '__NEXT_CLIENT_ROUTER_S_FILTER',
-  '__NEXT_LINK_NO_TOUCH_START',
-  '__NEXT_OPTIMISTIC_CLIENT_CACHE',
-  '__NEXT_SCROLL_RESTORATION',
-  '__NEXT_HAS_REWRITES',
-  '__NEXT_CROSS_ORIGIN',
-  '__NEXT_MANUAL_CLIENT_BASE_PATH',
-  '__NEXT_STRICT_NEXT_HEAD',
-  '__NEXT_TRAILING_SLASH',
-  '__NEXT_EXTERNAL_MIDDLEWARE_REWRITE_RESOLVE',
-  '__NEXT_CLIENT_ROUTER_FILTER_ENABLED',
-  '__NEXT_MIDDLEWARE_PREFETCH',
-  'NEXT_RUNTIME',
-  'NEXT_DEPLOYMENT_ID',
-]
+// We already strip most of the env variables in the built of the editor,
+// however React seems to reintroduce this one!
+const productionKeys = ['process.env.NODE_ENV']
 
 const envReplacements = {
-  ...Object.fromEntries(productionKeys.map((key) => [key, js('production')])),
   ...Object.fromEntries(
-    notProvidedKeys.map((key) => [
-      `process.env.${key}`,
-      js(`NOT_PROVIDED_${key}`),
-    ])
+    productionKeys.map((key) => [key, JSON.stringify('production')])
   ),
 }
