+++ packages/editor/package.json
@@ -68,6 +68,7 @@
     "@radix-ui/react-navigation-menu": "^1.1.4",
     "@radix-ui/react-select": "^1.2.2",
     "@reduxjs/toolkit": "^1.9.7",
+    "@rollup/plugin-replace": "^5.0.5",
     "@serlo/authorization": "^0.60.0",
     "@serlo/frontend": "workspace:*",
     "@serlo/typescript-config": "workspace:*",
